+++ imports/api/api.js
@@ -19,6 +19,7 @@ import './genomes/removeAnnotationTrack.js';
 import './genes/interproscan.js';
 import './genes/addInterproscan.js';
 import './genes/eggnog/addEggnog.js';
+import './genes/eggnog/addEggnog.js';
 import './genes/scanGeneAttributes.js';
 import './genes/updateAttributeInfo.js';
 import './genes/updateGene.js';