+++ CHANGELOG.md
@@ -7,9 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.3.2]
+
 ### Added
 
 - Support for additional InterproScan file formats (tsv, gff, xml)
+- EggNog annotations
+- BLAST/diamond alignment reading/visualizing
+
+### Changed
+
+- Meteor version 2.8.0
 
 ## [0.3.1] 2021-01-28
 
@@ -281,7 +289,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - First GeneNoteBook version ready for use
 
-[unreleased]: https://github.com/genenotebook/genenotebook/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/genenotebook/genenotebook/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/genenotebook/genenotebook/compare/v0.3.1...v0.3.2
+[0.3.1]: https://github.com/genenotebook/genenotebook/compare/v0.3.0...v0.3.1
+[0.3.0]: https://github.com/genenotebook/genenotebook/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/genenotebook/genenotebook/compare/v0.1.16...v0.2.0
 [0.1.16]: https://github.com/genenotebook/genenotebook/compare/v0.1.15...v0.1.16
 [0.1.15]: https://github.com/genenotebook/genenotebook/compare/v0.1.14...v0.1.15